Object definition for workflow queue item
Initial prototype code, to be populated from appropriate data feeds. Note that some properties are arrays to allow for the management of change history items (e.g. where priorities are amended etc) - these will need some further work to make sure that they are correctly defined and the information required is populated into them as and when required, and in a way that can be extracted for analysis later.